+++ Api.Application/Controllers/UserController.cs
@@ -8,21 +8,24 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Application.Controllers
 {
-    //http://localhost:5000/api/users
     [Route("api/[controller]")]
     [ApiController]
     public class UsersController : ControllerBase
     {
-        public IUserService _service { get; set; }
+        private IUserService _service { get; set; }
         public UsersController(IUserService service)
         {
             _service = service;
         }
 
-        [Authorize("Bearer")]
+       // [Authorize("Bearer")]
         [HttpGet]
         public async Task<ActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             try
             {
                 return Ok(await _service.GetAll());
@@ -33,11 +36,15 @@ namespace Api.Application.Controllers
             }
         }
 
-        [Authorize("Bearer")]
+       // [Authorize("Bearer")]
         [HttpGet]
         [Route("{id}", Name = "GetWithId")]
         public async Task<ActionResult> Get(Guid id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             try
             {
                 return Ok(await _service.Get(id));
@@ -49,10 +56,14 @@ namespace Api.Application.Controllers
         }
 
 
-        [Authorize("Bearer")]
+      //  [Authorize("Bearer")]
         [HttpPost]
         public async Task<ActionResult> Post([FromBody] UserDtoCreate category)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             try
             {
                 var result = await _service.Post(category);
@@ -69,10 +80,14 @@ namespace Api.Application.Controllers
             }
         }
 
-        [Authorize("Bearer")]
+     //   [Authorize("Bearer")]
         [HttpPut]
         public async Task<ActionResult> Put([FromBody] UserDtoUpdate user)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             try
             {
                 var result = await _service.Put(user);
@@ -91,10 +106,14 @@ namespace Api.Application.Controllers
             }
         }
 
-        [Authorize("Bearer")]
+     //   [Authorize("Bearer")]
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(Guid id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             try
             {
                 return Ok(await _service.Delete(id));
